requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-JWT-Extended
Django
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers 